refactor macro guard for composable kernel; refactor some host code
