+++ src/kernels/composable_kernel/include/utility/amd_buffer_addressing.hpp
@@ -138,7 +138,7 @@ __llvm_amdgcn_buffer_store_bf16x4(ushort4_t vdata,
                                   bool glc,
                                   bool slc) __asm("llvm.amdgcn.buffer.store.v4bf16");
 
-#if CK_USE_AMD_BUFFER_ATOMIC_ADD
+#if CK_HARDWARE_SUPPORT_AMD_BUFFER_ATOMIC_FADD
 __device__ void
 __llvm_amdgcn_buffer_atomic_add_f32(float vdata,
                                     int32x4_t rsrc,
@@ -165,7 +165,7 @@ __device__ void amd_buffer_store(const T* p_src,
                                  index_t dst_thread_data_offset,
                                  index_t dst_const_data_offset);
 
-#if CK_USE_AMD_BUFFER_ATOMIC_ADD
+#if CK_HARDWARE_SUPPORT_AMD_BUFFER_ATOMIC_FADD
 template <typename T, index_t VectorSize>
 __device__ void amd_buffer_atomic_add(const T* p_src,
                                       T* p_dst_block,
@@ -738,7 +738,7 @@ __device__ void amd_buffer_store<ushort, 4>(const ushort* p_src,
 #endif
 }
 
-#if CK_USE_AMD_BUFFER_ATOMIC_ADD
+#if CK_HARDWARE_SUPPORT_AMD_BUFFER_ATOMIC_FADD
 template <>
 __device__ void amd_buffer_atomic_add<float, 1>(const float* p_src,
                                                 float* p_dst_block,
@@ -786,7 +786,7 @@ __device__ void amd_buffer_atomic_add<float, 4>(const float* p_src,
             &p_src[i], p_dst_block, dst_thread_data_offset, dst_const_data_offset + i);
     }
 }
-#endif // CK_USE_AMD_BUFFER_ATOMIC_ADD
+#endif // CK_HARDWARE_SUPPORT_AMD_BUFFER_ATOMIC_FADD
 
 } // namespace ck
 #endif
