+++ src/solver/conv_hip_implicit_gemm_v4r4.cpp
@@ -23,14 +23,14 @@
  * SOFTWARE.
  *
  *******************************************************************************/
+#include <cstddef>
 #include <miopen/conv/invokers/impl_gemm.hpp>
 #include <miopen/solver.hpp>
 #include <miopen/handle.hpp>
 #include <miopen/generic_search.hpp>
-
 #include "implicitgemm_util.hpp"
-
-#include <cstddef>
+// TODO remove dependency on deprecated header
+#include "implicitgemm_util_deprecated.hpp"
 
 namespace miopen {
 namespace solver {
@@ -758,6 +758,7 @@ ConvSolution ConvHipImplicitGemmV4R4Fwd::GetSolution(const ConvolutionContext& c
         std::string(" -DCK_PARAM_DEPENDENT_GRID_SIZE=") + std::to_string(grid_size) +
         std::string(" -DCK_THREADWISE_GEMM_USE_AMD_INLINE_ASM=") + (use_amd_inline_asm(ctx) ? '1' : '0') +
         std::string(" -DCK_USE_AMD_INLINE_ASM=") + (use_amd_inline_asm(ctx) ? '1' : '0') +
+        get_compiler_flag_for_supported_instruction(ctx) +
         ctx.general_compile_options;
 
         if (ctx.Is3d()){
