+++ src/solver/conv_hip_implicit_gemm_v4r4_gen_xdlops.cpp
@@ -25,7 +25,6 @@
  *******************************************************************************/
 
 #include <miopen/solver.hpp>
-
 #include <miopen/conv/invokers/impl_gemm.hpp>
 #include <miopen/conv/wrw_invoke_params.hpp>
 #include <miopen/handle.hpp>
@@ -33,8 +32,9 @@
 #include <miopen/stringutils.hpp>
 #include <miopen/tensor_ops.hpp>
 #include <miopen/implicitgemm_params.hpp>
-
 #include "implicitgemm_util.hpp"
+// TODO remove dependency on deprecated header
+#include "implicitgemm_util_deprecated.hpp"
 
 namespace miopen {
 namespace solver {
@@ -294,6 +294,7 @@ static inline ConvSolution GetSolutionBase(const ConvolutionContext& ctx,
         std::string(" -DCK_USE_AMD_XDLOPS=") + std::to_string(IsXdlopsSupport(ctx) ? 1 : 0) +
         std::string(" -DCK_USE_AMD_XDLOPS_INLINE_ASM=") + std::to_string(miopen::IsEnabled(MIOPEN_DEBUG_IMPLICIT_GEMM_XDLOPS_INLINE_ASM{}) ? 1 : 0) +
         std::string(" -DCK_USE_AMD_XDLOPS_EMULATE=") + (miopen::IsEnabled(MIOPEN_DEBUG_CONV_IMPLICIT_GEMM_XDLOPS_EMULATE{}) ? '1' : '0') +
+        get_compiler_flag_for_supported_instruction(ctx) +
         ctx.general_compile_options;
     // clang-format on
